now it should pass all the tests...now it's for real. Float everywhere

+++ src/main/java/com/packtpub/Calculator.java
@@ -9,7 +9,7 @@ public class Calculator {
         return x + y;
     };
 
-    public float div(int x, int y)
+    public float div(float x, float y)
     {
         return x/y;
     }
